+++ app/assets/stylesheets/application.scss.erb
@@ -3150,11 +3150,8 @@ script.data-gratipay-username {
 }
 
 .topicFollow {
-    text-align: center;
     height: 48px;
     line-height: 48px;
-    border-top: 1px solid #BDBDBD;
-    background: #FFF;
     cursor: pointer;
     font-family: din-regular;
 }
